pyproject.toml
[project]
name = "friendly-computing-machine"
version = "0.1.0"
description = "slackbot"
readme = "README.md"
authors = [
    { name = "qmastermoo", email = "alex.harding@whale-net.net" }
]
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0",
    "google-generativeai>=0.8.4",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-distro>=0.50b0",
    "opentelemetry-exporter-otlp>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.3",
    "python-dateutil>=2.9.0.post0",
    "slack-bolt>=1.21.2",
    "sqlalchemy>=2.0.36",
    "sqlmodel>=0.0.22",
    "temporalio[opentelemetry]>=1.10.0",
    "typer-slim>=0.15.2",
]

[project.scripts]
fcm = "friendly_computing_machine:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
# sources = ["."]

[tool.pytest.ini_options]
minversion = "6.0"
# addopts = "-ra -q"
testpaths = [
    "tests",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.6.1",
]
